feat(context, frontend): Enhance CONTEXT.md, README.md, and implement Asset Tab functionality
- Updated CONTEXT.md to include a new section detailing the frontend implementation of the Asset Tab, including features like simulation parameters, results dashboard, and API integration with the FastAPI backend.
- Enhanced README.md to reflect the functionality of the Asset Tab, allowing users to select assets, configure DCA parameters, and view interactive charts.
- Refactored the simulation engine to ensure proper type hinting and data validation, improving overall performance and reliability.
- Updated Tailwind CSS configuration to include new color and font settings for better UI consistency.
- Made various adjustments to layout and page components to support the new features and improve user experience.

+++ backend/engine/simulation_engine.py
@@ -122,15 +122,15 @@ class SimulationEngine:
                 "invested": round(total_invested, 2)
             })
 
-        final_value = assets_accumulated * self.market_data.iloc[-1]['close']
-        total_return_percent = ((final_value - total_invested) / total_invested * 100) if total_invested > 0 else 0
-        avg_price = total_invested / assets_accumulated if assets_accumulated > 0 else 0
+        final_value = float(assets_accumulated * self.market_data.iloc[-1]['close'])
+        total_return_percent = float(((final_value - total_invested) / total_invested * 100)) if total_invested > 0 else 0.0
+        avg_price = float(total_invested / assets_accumulated) if assets_accumulated > 0 else 0.0
 
         return DCAResult(
             portfolio_history=portfolio_history,
             final_value=round(final_value, 2),
-            total_invested=round(total_invested, 2),
+            total_invested=round(float(total_invested), 2),
             total_return_percent=round(total_return_percent, 2),
             avg_purchase_price=round(avg_price, 2),
-            total_assets_accumulated=assets_accumulated
+            total_assets_accumulated=float(assets_accumulated)
         )
